test the completion of data from r2c_1d dft
The real to complex DFT routines of FFTW do not compute all the output,
but instead compute only half the values, since the remaining half are
the complex conjugates of the first half.

A test was written to compute the FFT of real data, complete it, then
use the completed data in an inverse FFT to obtain the real data again.
The test ensures that the sup-norm of the vector is small, thus the
implementation is correct.